+++ src/primitives/Array.cpp
@@ -54,7 +54,7 @@ _Array::_Array() :
 void _Array::At_(pVMObject /*object*/, pVMFrame frame) {
     pVMInteger index = static_cast<pVMInteger>(frame->Pop());
     pVMArray self = static_cast<pVMArray>(frame->Pop());
-    int i = index->GetEmbeddedInteger();
+    long i = index->GetEmbeddedInteger();
     pVMObject elem = (*self)[i-1];
     frame->Push(elem);
 }
@@ -63,7 +63,7 @@ void _Array::At_Put_(pVMObject /*object*/, pVMFrame frame) {
     pVMObject value = frame->Pop();
     pVMInteger index = static_cast<pVMInteger>(frame->Pop());
     pVMArray self = static_cast<pVMArray>(frame->GetStackElement(0));
-    int i = index->GetEmbeddedInteger();
+    long i = index->GetEmbeddedInteger();
     (*self)[i - 1] = value;
 }
 
@@ -77,7 +77,7 @@ void _Array::Length(pVMObject /*object*/, pVMFrame frame) {
 void _Array::New_(pVMObject /*object*/, pVMFrame frame) {
     pVMInteger length = static_cast<pVMInteger>(frame->Pop());
     frame->Pop();
-    int size = length->GetEmbeddedInteger();
+    long size = length->GetEmbeddedInteger();
     frame->Push(_UNIVERSE->NewArray(size));
 }
